venture_engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
apscheduler==3.10.4
anthropic==0.42.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
python-dotenv==1.0.1
jinja2==3.1.4
loguru==0.7.3
tenacity==9.0.0
aiosqlite==0.20.0
pydantic==2.10.4
pydantic-settings==2.7.1
